LayoutTests/fast/sub-pixel/block-preferred-widths-with-sub-pixel-floats.html

<html>
<head>
<script src="../../fast/js/resources/js-test-pre.js"></script>
<style type="text/css">
#relativeContainer { position:relative; width:980px; height:58px; margin:18px auto 0; z-index:999; font-size:12px; }
#relativeContainer ul { position:absolute; right:200px; top:1px;  margin:0; }
#relativeContainer ul li { display:inline; float:left; margin:0 0 0 2.8em; background-color: green;}
#relativeContainer ul li a { display:block; color:#333; text-decoration:none; padding:.3em 0 .3em 5px; text-shadow: 0 1px 0 #fff; line-height:18px; }
</style>	
</head>
<body>
<div id="relativeContainer">
  <ul>
    <li id="support"><a href="#">Support</a></li>
    <li id="membercenter"><a href="#">Member Center</a></li>
 </ul>
</div>
<p id="description"></p>
<div id="console"></div>
<script>
description("This tests that we don't cause an assertion failure on relayout of nested positioned elements. This test pass if we don't cause an assertion failure.");

shouldBe('document.getElementById("support").offsetTop', 'document.getElementById("membercenter").offsetTop');

if (window.testRunner)
	document.getElementById("relativeContainer").display = 'none';

</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
