third_party/blink/web_tests/fast/sub-pixel/block-preferred-widths-with-sub-pixel-floats-expected.txt

Support
Member Center

This tests that we don't cause an assertion failure on relayout of nested positioned elements. This test pass if we don't cause an assertion failure.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.getElementById("support").offsetTop is document.getElementById("membercenter").offsetTop
PASS successfullyParsed is true

TEST COMPLETE
